add configuracion y auditoria

+++ spring-jsf/src/main/java/pe/edu/cibertec/core/service/impl/LaboratorioServiceImpl.java
@@ -1,7 +1,6 @@
 package pe.edu.cibertec.core.service.impl;
 
 import java.util.List;
-import java.util.jar.Attributes.Name;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,30 +12,40 @@ import org.springframework.transaction.annotation.Transactional;
 import pe.edu.cibertec.core.dao.LaboratorioDAO;
 import pe.edu.cibertec.core.domain.Laboratorio;
 import pe.edu.cibertec.core.service.LaboratorioService;
-@Service("laboratorioService")
+import pe.edu.cibertec.exception.BusinessException;
+@Service
 @Transactional
 public class LaboratorioServiceImpl implements LaboratorioService {
 
 	private static final Logger logger=LoggerFactory.getLogger(LaboratorioServiceImpl.class);
+	
 	@Autowired
 	private LaboratorioDAO laboratorioDAO;
 	@Override
-	@Transactional(readOnly=false,propagation=Propagation.REQUIRED,rollbackFor=Exception.class)
-	public void save(Laboratorio laboratorio) {
+	@Transactional(readOnly=false,propagation=Propagation.REQUIRED,rollbackFor=BusinessException.class)
+	public void save(Laboratorio laboratorio)  throws BusinessException{
 		logger.info("save");
-		laboratorioDAO.save(laboratorio);
+		if(laboratorio.getLaboratorioId()!=null){
+			throw new BusinessException(BusinessException.SEVERITY_CRITICAL, "Estas intentando Actualizar cuando debe Registar un Nuevo Laboratorio");
+		}else{			
+			laboratorioDAO.save(laboratorio);
+		}
 		
 	}
 
 	@Override
-	@Transactional(readOnly=false,propagation=Propagation.REQUIRED,rollbackFor=Exception.class)
-	public void update(Laboratorio laboratorio) {
-		laboratorioDAO.update(laboratorio);
+	@Transactional(readOnly=false,propagation=Propagation.REQUIRED,rollbackFor=BusinessException.class)
+	public void update(Laboratorio laboratorio) throws BusinessException {
+		if(laboratorio.getLaboratorioId()!=null){
+			laboratorioDAO.update(laboratorio);
+		}else{			
+			throw new BusinessException(BusinessException.SEVERITY_NORMAL, "Por favor verifica los datos..");
+		}
 		
 	}
 
 	@Override
-	public Laboratorio get(Laboratorio laboratorio) {		
+	public Laboratorio getById(Laboratorio laboratorio) {		
 		return laboratorioDAO.get(laboratorio);
 	}
 
@@ -46,9 +55,14 @@ public class LaboratorioServiceImpl implements LaboratorioService {
 	}
 
 	@Override
-	@Transactional(readOnly=false,propagation=Propagation.REQUIRED,rollbackFor=Exception.class)
+	@Transactional(readOnly=false,propagation=Propagation.REQUIRED,rollbackFor=BusinessException.class)
 	public void delete(Laboratorio laboratorio) {
-		laboratorioDAO.delete(laboratorio);
+
+		if(laboratorio.getLaboratorioId()!=null){
+			laboratorioDAO.delete(laboratorio);
+		}else{			
+			throw new BusinessException(BusinessException.SEVERITY_NORMAL, "Selecione bien el registro a dar de baja.");
+		}
 		
 	}
